add adding products page

+++ cart.php
@@ -11,7 +11,7 @@
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
 
     <!-- custom css file link  -->
-    <link rel="stylesheet" href="css/style.css">
+    <link rel="stylesheet" href="CSS/styleUser.css">
 
 </head>
 
@@ -34,24 +34,23 @@
          if(mysqli_num_rows($select_cart) > 0){
             while($fetch_cart = mysqli_fetch_assoc($select_cart)){   
       ?>
-            <div class="box">
-                <a href="cart.html?delete=<?php echo $fetch_cart['id']; ?>" class="fas fa-times"
-                    onclick="return confirm('delete this from cart?');"></a>
-                <img src="uploaded_img/<?php echo $fetch_cart['image']; ?>" alt="">
-                <div class="name">
-                    <?php echo $fetch_cart['name']; ?>
-                </div>
-                <div class="price">$
-                    <?php echo $fetch_cart['price']; ?>/-
-                </div>
-                <form action="" method="post">
-                    <input type="hidden" name="cart_id" value="<?php echo $fetch_cart['id']; ?>">
-                    <input type="number" min="1" name="cart_quantity" value="<?php echo $fetch_cart['quantity']; ?>">
-                    <input type="submit" name="update_cart" value="update" class="option-btn">
-                </form>
+                <div class="box">
+                    <a href="cart.html?delete=<?php echo $fetch_cart['id']; ?>" class="fas fa-times" onclick="return confirm('delete this from cart?');"></a>
+                    <img src="uploaded_img/<?php echo $fetch_cart['image']; ?>" alt="">
+                    <div class="name">
+                        <?php echo $fetch_cart['name']; ?>
+                    </div>
+                    <div class="price">$
+                        <?php echo $fetch_cart['price']; ?>/-
+                    </div>
+                    <form action="" method="post">
+                        <input type="hidden" name="cart_id" value="<?php echo $fetch_cart['id']; ?>">
+                        <input type="number" min="1" name="cart_quantity" value="<?php echo $fetch_cart['quantity']; ?>">
+                        <input type="submit" name="update_cart" value="update" class="option-btn">
+                    </form>
 
-            </div>
-            <?php
+                </div>
+                <?php
       $grand_total += $sub_total;
          }
       }else{
@@ -61,8 +60,7 @@
         </div>
 
         <div style="margin-top: 2rem; text-align:center;">
-            <a href="cart.html?delete_all" class="delete-btn <?php echo ($grand_total > 1)?'':'disabled'; ?>"
-                onclick="return confirm('delete all from cart?');">delete all</a>
+            <a href="cart.html?delete_all" class="delete-btn <?php echo ($grand_total > 1)?'':'disabled'; ?>" onclick="return confirm('delete all from cart?');">delete all</a>
         </div>
 
         <div class="cart-total">
